+++ exporter/opentelemetry-exporter-jaeger-thrift/examples/jaeger_exporter_example.py
@@ -1,7 +1,7 @@
 import time
 
 from opentelemetry import trace
-from opentelemetry.exporter import jaeger
+from opentelemetry.exporter.jaeger import thrift
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import BatchSpanProcessor
 
@@ -9,29 +9,15 @@ trace.set_tracer_provider(TracerProvider())
 tracer = trace.get_tracer(__name__)
 
 # create a JaegerExporter
-jaeger_exporter = jaeger.JaegerExporter(
+jaeger_exporter = thrift.JaegerExporter(
     # configure agent
     agent_host_name="localhost",
     agent_port=6831,
     # optional: configure also collector
-    # collector_host_name="localhost",
-    # collector_port=14268,
-    # collector_endpoint="/api/traces?format=jaeger.thrift",
+    # collector_endpoint="http://localhost:14268/api/traces?format=jaeger.thrift",
     # username=xxxx, # optional
     # password=xxxx, # optional
 )
-
-# Create a JaegerExporter to send spans with gRPC
-# If there is no encryption or authentication set `insecure` to True
-# If server has authentication with SSL/TLS you can set the
-# parameter credentials=ChannelCredentials(...) or the environment variable
-# `EXPORTER_JAEGER_CERTIFICATE` with file containing creds.
-
-# jaeger_exporter = jaeger.JaegerExporter(
-#     collector_endpoint="localhost:14250",
-#     insecure=True,
-#     transport_format="protobuf",
-# )
 
 # create a BatchSpanProcessor and add the exporter to it
 span_processor = BatchSpanProcessor(jaeger_exporter)